Error

Description

Types for errors that may be thrown when processing the user input. Enables for pretty display of errors using ParseErrorBundle.

Documentation

type CustomBundle = ParseErrorBundle String Error

ParseErrorBundle instantiated with the custom type Error.

data ErrorType

Types of errors that may be thrown.

Constructors

ReservedError [String]
Error for user trying to define a reserved name.

ModuleDuplicationError [String]
Error for user defining a module multiple times.

FindModuleError [String]
Error for when a module can not be found.

UndefinedError
Error for when a name is undefined.

ShadowingError [String]
Error for when an expression reintroduces a name.

TopologicalError
Error for when a topological sort is impossible.

TypeError
Error for when a valid type could not be deduced for a declaration.

data Error

An ErrorType alongside associated information.

Constructors

Error

Fields

errType :: ErrorType

errLen :: Int

The length of the parsed item for which the error is being thrown, which is used for pretty printing.

mkParseError

Arguments

:: ErrorType

-> LocAnn x
The location where the error resides.

-> ParseError String Error

Makes a ParseError.

mkBundleError

Arguments

:: ErrorType

-> LocAnn x
The location where the error resides.

-> CustomBundle

Makes a CustomBundle.