ModuleResolver

Description

Processes import statements within files to convert a list of FileContents to a global list of MainStatements, with names of constants being replaced by unique integers.

Documentation

data UniqueNames

Type used to give each constant a name that is globally unique.

Constructors

UniqueName Integer
An integer value for a constant that is guaranteed to be unique.

BoundVariable [String]
The name of a variable bound within an expression. The converter errors if a bound name is not unique within the scope of the expression.

Validity checks

findDupBy :: Ord b => (a -> b) -> [a] -> Maybe [a]

Gives a list of at least two elements for which the key function gives the same value.

checkDuplicateModule :: [FileContents] -> Either CustomBundle ()

Ensures that no two files have the same module name. Throws ModuleDuplicationError otherwise.

checkShadowing

Arguments

:: Map [String] a
A map containing the defined names.

-> LocAnn x
The location where the name is introduced.

-> [String]
The name being introduced.

-> Either CustomBundle ()

Ensures that the introduction of a name does not overlap with a previously defined name. Throws ShadowingError otherwise.

Conversion functions.

assignNamesState :: [FileContents] -> State Integer [([String], [([String], Integer)])]

Given a state, assigns, for each name in each module, a globally unique integer name.

makeNameMapping

Arguments

:: [FileContents]

-> Map [String] [([String], Integer)]
A mapping from module name to the names within that module and their associated unique integers.

Assigns unique names to each name in each file.

introduceImport

Arguments

:: Map [String] [([String], Integer)]
A mapping from module names to names within that module and their unique integer assignments.

-> Map [String] Integer
A mapping containing names already defined within the scope.

-> LocAnn ImportStatement

-> Either CustomBundle (Map [String] Integer)

Introduces all the qualified names from an import, and checks these names do not overlap previously defined names. Throws FindModuleError if a module can not be found.

convertExpr

Arguments

:: Map [String] Integer
The file namespace: a mapping from names to unique integers.

-> Expr [String]

-> Either CustomBundle (Expr UniqueNames)

Converts the names within an expression to unique names.

convertStatement

Arguments

:: Map [String] Integer
The file namespace: a mapping from names to unique integers.

-> MainStatement [String]

-> Either CustomBundle (MainStatement UniqueNames)

Converts the names within a main statement to unique integers.

moduleToMainStatements

Arguments

:: Map [String] [([String], Integer)]
The result of makeNameMapping on all file contents.

-> FileContents

-> Either CustomBundle [MainStatement UniqueNames]

Converts the names within a module to unique integers.

toMainStatements :: [FileContents] -> Either CustomBundle [MainStatement UniqueNames]

Converts a list of FileContents to a global list of MainStatement UniqueNames.