Parser

Description

Parsers that are ultimately used to form a list of FileContents from a directory.

Documentation

type Parser = Parsec Error String

The type used for the language parser.

spaceConsumer :: Parser ()

Consumes whitespace, line comments beginning with --, and block comments, which have start indicated by {- and end indicated by -}.

spaceConsumer' :: Parser ()

Consumes horizontal whitespace, line comments beginning with --, and block comments, which have start indicated by {- and end indicated by -}.

lexeme :: Parser a -> Parser a

Consumes whitespace using spaceConsumer after running a parser.

lexeme' :: Parser a -> Parser a

Consumes whitespace using 'spaceConsumer'' after running a parser.

lexemeS :: Parser a -> Parser a

Run a parser and then consume whitespace using spaceConsumer. Also ensures that the string is followed by either whitespace, or one of the characters ",:;=()".

symbol :: String -> Parser String

Parses a string and then consume whitespace using spaceConsumer.

symbol' :: String -> Parser String

Parses a string and then consume whitespace using 'spaceConsumer''.

symbolS :: String -> Parser String

Parses a string using lexemeS.

errAnn :: Parser x -> Parser (LocAnn x)

Annotates a parsed value with LocAnn.

reservedIdentifiers :: [[String]]

Names which can not be defined within the language.

identifierParser :: Parser (LocAnn [String])

Parses an identifier, which consists of a list of strings (consisting of characters that are either alphanumeric or one of _+*/-~$!) of length at least 1, and also ensures this identifier is not reserved.

identifierParser' :: Parser [String]

'identifierParser'' without an annotation.

expressionParser1 :: Parser (Expr [String])

Parses expressions within the language, with the exception of top level applications. Such expressions are parsed separately to ensure left associativity.

expressionParser :: Parser (Expr [String])

Parses top level applications and combines them in a left recursive manner.

moduleNameParser :: Parser [String]

Parses a declaration of the module name.

importParser :: Parser ImportStatement

Parses an import statement, which is either of the form import module as qualifier or import module.

mainStatementParser :: Parser (MainStatement [String])

Parses the main statements within a file. All assumptions are of the form x : T, with definitions being either of the form x = e or x : T = e.

fileContentsParser :: Parser FileContents

Parses the module name, import statements, and main statements within a file.

getFiles :: String -> IO [String]

Give names of files in a directory (relative to that directory) in the data directory. All names are given relative to this root directory, and the presence of symbolic links within this root directory will lead to an error being thrown.

input :: String -> ExceptT CustomBundle IO [FileContents]

Parses all the file contents within a root directory within the data directory.