Syntax

Description

Data types representing the syntax of the programming language, as well as data types for annotating source locations.

Documentation

data Expr x

An expression in the language.

Constructors

Set
The Set type.

Prop
The Prop type.

Type Integer
The Type(i) type. The types are 0 indexed.

Identifier (LocAnn x)
A variable or constant in the language.

Forall
The expression forall x : T, U.

Fields

(LocAnn x)

The name of the variable bound in U.

(Expr x)

The type T of the variable bound.

(Expr x)

The type U.

Abs
The expression fun x : T, u.

Fields

(LocAnn x)

The name of the variable bound in u.

(Expr x)

The type T of the variable bound.

(Expr x)

The expression u.

Let
The expression let x : T = t in u or let x = t in u.

Fields

(LocAnn x)

The name of the variable bound in u.

(Maybe (Expr x))

The type T.

(Expr x)

The expression t.

(Expr x)

The expression u.

App (Expr x) (Expr x)
An application of two expressions.

data LocAnn x

An annotation to help throwing errors at a specific code location.

Constructors

LocAnn

Fields

annPosState :: PosState String

State stored to convert offsets to inputs.

annOffset :: Int

The offset where parsing of the item began.

annLen :: Int

The length of the item parsed.

annVal :: x

NoLoc

Fields

annVal :: x

data ImportStatement

A statement to introduce identifiers from other modules into the namespace.

Constructors

Import
The statement import module as name or import module.

Fields

importName :: [String]

The name of the module to import.

importAs :: [String]

Potentially empty list of qualifiers to prepend to the identifiers imported.

data MainStatement x

The main statements in the file.

Constructors

Assume
A type assumption x : T.

Fields

(LocAnn x)

The variable x.

(Expr x)

The type T.

Define
A declaration, of form either x = y or, to verify the type of the expression, x : T = y.

Fields

(LocAnn x)

The variable x.

(Maybe (Expr x))

The optional type T.

(Expr x)

The expression y

extractAnnDef :: MainStatement x -> LocAnn x

Get the variable being defined by a MainStatement, alongside its LocAnn.

extractDef :: MainStatement x -> x

Get the variable being defined by a MainStatement.

data FileContents

The information contained in a file.

Constructors

FileContents

Fields

moduleName :: LocAnn [String]

The name of the module to associate the file to.

importStmts :: [LocAnn ImportStatement]

The imports within the file.

mainStmts :: [MainStatement [String]]

The main statements within the file.

showName :: [String] -> String

Displays an identifier by combining the qualifiers with full stops.