TypeChecking

Description

Ensures that terms defined in language are well-formed and match their type annotations.

Documentation

type ExprMap = Map Integer (Expr ResolvedName)

Type synonym for maps to expressions.

Ordering statements

topologicalSort :: [MainStatement ResolvedName] -> Either CustomBundle [MainStatement ResolvedName]

Creates, if possible, an ordering of the MainStatements such that no MainStatement depends on a name not already defined. Throws TopologicalError otherwise.

Normalisation of the lambda calculus

applyShift

Arguments

:: Integer
The level at which the shift starts.

-> Integer
The shift to apply.

-> ResolvedName

-> ResolvedName

Increases an index by a shift if its value is greater than or equal to a given index index.

replaceLvl

Arguments

:: Integer
The shift to apply to bound terms within the replacement term. This is determined by the number of additional bindings that come before the term.

-> Integer
The level of the binding x.

-> Expr ResolvedName
The replacement term a.

-> Expr ResolvedName
The term u.

-> Expr ResolvedName

Calculates the application (fun x : T => u) a, by decrementing all indices greater in level than the binding x, and replacing the binding x with the term a, with the bound levels within this term being increased as appropriate.

normaliseExpr

Arguments

:: Integer
The current depth of the expression. This is the level which would be given to a new binding in the subexpression.

-> ExprMap
A mapping from the unique integer given to a global name to the expression for that name.

-> Expr ResolvedName

-> Expr ResolvedName

Calculates the normal form of an expression.

Type checking

isSort :: Expr x -> Bool

Determines whether an expression is a sort.

areEqual :: Expr ResolvedName -> Expr ResolvedName -> Bool

Determines whether two expressions are equal when location annotations are removed.

areEtaEqual

Arguments

:: Integer
The depth at which the comparison is being done.

-> Expr ResolvedName

-> Expr ResolvedName

-> Bool

Determines whether two expressions are equal up to eta expansion.

isConvertible :: Expr ResolvedName -> Expr ResolvedName -> Bool

Determines whether an expression is convertible to another according to the subtyping relationship.

checkConvertible

Arguments

:: LocAnn x

-> Integer

-> ExprMap
A mapping from the unique integer given to a global name to the expression for that name.

-> Expr ResolvedName

-> Expr ResolvedName

-> Either CustomBundle ()

Guards on the convertibility of two expressions.

formType

Arguments

:: LocAnn x

-> Integer
The subexpression depth.

-> ExprMap
A mapping from De Bruijn levels to the type of that binding.

-> ExprMap
A mapping from the unique integer for each global name to the value of that expression.

-> ExprMap
A mapping from the unique integer for each global name to the type of that expression.

-> Expr ResolvedName

-> Either CustomBundle (Expr ResolvedName)

Forms a type for an expression or subexpression.

introTypes :: [MainStatement ResolvedName] -> Either CustomBundle (ExprMap, ExprMap)

Introduces each statement into the global context according to the rules for well-founded global contexts.